requirements.txt
requests
tk
pywebview
sv-ttk